fix(search): reject fractional HTTP result limits

+++ apps/backend/src/routes/api/hypermedia/search/model.ts
@@ -36,7 +36,7 @@ export const HypermediaSearchQuerySchema = t.Object({
     }),
   ),
   limit: t.Optional(
-    t.Numeric({
+    t.Integer({
       minimum: 1,
       maximum: MAX_HYPERMEDIA_SEARCH_LIMIT,
       default: DEFAULT_HYPERMEDIA_SEARCH_LIMIT,
